MODIFIED: Fixed calc_num_block
Fixed a bug where an integer divide by an integer would return a integer value rather then a double.

+++ src/block_func.c
@@ -47,5 +47,5 @@ uint32_t get_num_datablocks(byte* data_buf)
 	{
 		i++;
 	}
-	return (uint32_t)ceil(i/BLKSIZE);
+	return (uint32_t)ceil((double)(i/BLKSIZE));
 }
